+++ scripts/actions/alchemical-shot.js
@@ -1,6 +1,9 @@
+import { Weapon } from "../types/pf2e-ranged-combat/weapon.js";
+import { PF2eActor } from "../types/pf2e/actor.js";
+import { PF2eItem } from "../types/pf2e/item.js";
 import { HookManager } from "../utils/hook-manager.js";
 import { Updates } from "../utils/updates.js";
-import { ensureDuration, getControlledActorAndToken, getEffectFromActor, getFlag, getItem, getItemFromActor, postActionToChat, postToChat, setEffectTarget, showWarning } from "../utils/utils.js";
+import { ensureDuration, getControlledActor, getEffectFromActor, getFlag, getItem, getItemFromActor, postActionToChat, postToChat, setEffectTarget, showWarning } from "../utils/utils.js";
 import { getWeapon } from "../utils/weapon-utils.js";
 
 const ALCHEMICAL_SHOT_FEAT_ID = "Compendium.pf2e.feats-srd.Item.Q1O4P1YIkCfeedHH";
@@ -9,41 +12,84 @@ const ALCHEMICAL_SHOT_EFFECT_ID = "Compendium.pf2e-ranged-combat.effects.Item.IY
 const localize = (key) => game.i18n.localize("pf2e-ranged-combat.actions.alchemicalShot." + key);
 const format = (key, data) => game.i18n.format("pf2e-ranged-combat.actions.alchemicalShot." + key, data);
 
-export function initialiseAlchemicalShort() {
+export function initialiseAlchemicalShot() {
     HookManager.register("weapon-attack", handleWeaponFired);
+    HookManager.register("weapon-damage", handleWeaponDamage);
+    HookManager.register("post-action", handlePostAction);
+}
+
+/**
+ * Handle the action being posted to chat.
+ * 
+ * @param {{ actor: PF2eActor, item: PF2eItem, result: {match: boolean} }} params
+ */
+function handlePostAction({ actor, item, result }) {
+    if (item.sourceId != ALCHEMICAL_SHOT_FEAT_ID) {
+        return;
+    }
+
+    result.match = true;
+
+    performForActor(actor, item);
 }
 
 export async function alchemicalShot() {
-    const { actor, token } = getControlledActorAndToken();
+    const actor = getControlledActor();
     if (!actor) {
         return;
     }
 
     const alchemicalShotFeat = getItemFromActor(actor, ALCHEMICAL_SHOT_FEAT_ID);
     if (!alchemicalShotFeat) {
-        showWarning(format("warningNoFeat", { token: token.name }));
+        showWarning(format("warningNoFeat", { actor: actor.name }));
         return;
     }
 
+    performForActor(actor, alchemicalShotFeat);
+}
+
+/**
+ * @param {PF2eActor} actor 
+ * @param {PF2eItem} action
+ */
+async function performForActor(actor, action) {
     const weapon = await getWeapon(
         actor,
         weapon => weapon.isEquipped && (weapon.group == "firearm" || weapon.group == "crossbow"),
-        format("warningNotWieldingProperWeapon", { token: token.name })
+        format("warningNotWieldingProperWeapon", { actor: actor.name })
     );
     if (!weapon) {
         return;
     }
 
+    performForActorAndWeapon(actor, action, weapon);
+}
+
+/**
+ * @param {PF2eActor} actor
+ * @param {PF2eItem} action
+ * @param {Weapon} weapon
+ */
+async function performForActorAndWeapon(actor, action, weapon) {
     const bomb = await getWeapon(
         actor,
         weapon =>
             weapon.baseType === "alchemical-bomb"
             && weapon.quantity > 0,
-        format("warningNoAlchemicalBombs", { token: token.name })
+        format("warningNoAlchemicalBombs", { actor: actor.name })
     );
     if (!bomb) {
         return;
     }
+
+    await postActionToChat(action);
+    await postToChat(
+        actor,
+        bomb.img,
+        format("pourBombIntoWeapon", { actor: actor.name, weapon: weapon.name, bomb: bomb.name }),
+        localize("alchemicalShot"),
+        2
+    );
 
     const updates = new Updates(actor);
 
@@ -86,15 +132,6 @@ export async function alchemicalShot() {
         }
     );
 
-    await postActionToChat(alchemicalShotFeat);
-    await postToChat(
-        actor,
-        bomb.img,
-        format("tokenPoursBombIntoWeapon", { token: token.name, weapon: weapon.name, bomb: bomb.name }),
-        localize("alchemicalShot"),
-        2
-    );
-
     updates.handleUpdates();
 }
 
@@ -118,5 +155,17 @@ function handleWeaponFired({ weapon, updates }) {
                 }
             );
         }
+    }
+}
+
+/**
+ * Remove any Alchemical Shot effects for the weapon.
+ * 
+ * @param {{ weapon: Weapon, updates: Updates }}
+ */
+function handleWeaponDamage({ weapon, updates }) {
+    const alchemicalShotEffect = getEffectFromActor(weapon.actor, ALCHEMICAL_SHOT_EFFECT_ID, weapon.id);
+    if (alchemicalShotEffect) {
+        updates.delete(alchemicalShotEffect);
     }
 }