+++ scripts/ammunition-system/actions/reload.js
@@ -332,7 +332,7 @@ export async function performReload(actor, token, weapon, updates, options = {})
         await postReloadToChat(token, weapon);
     }
 
-    await HookManager.call("reload", { weapon, updates });
+    HookManager.call("reload", { weapon, updates });
 
     Hooks.callAll("pf2eRangedCombatReload", actor, token, weapon);
 };